move operators into hcl-primitives
