+++ crates/hcl-rs/src/expr/operation.rs
@@ -1,8 +1,9 @@
 use super::Expression;
-use crate::{Error, Result};
 use serde::Deserialize;
-use std::fmt;
-use std::str::FromStr;
+
+// Re-exported for convenience.
+#[doc(inline)]
+pub use hcl_primitives::expr::{BinaryOperator, UnaryOperator};
 
 /// Operations apply a particular operator to either one or two expression terms.
 #[derive(Deserialize, Debug, PartialEq, Eq, Clone)]
@@ -43,43 +44,6 @@ impl UnaryOp {
         UnaryOp {
             operator,
             expr: expr.into(),
-        }
-    }
-}
-
-/// An operator that can be applied to an expression.
-#[derive(Debug, PartialEq, Eq, Clone, Copy)]
-pub enum UnaryOperator {
-    /// Negate operator (`-`).
-    Neg,
-    /// Not operator (`!`).
-    Not,
-}
-
-impl UnaryOperator {
-    /// Returns the `UnaryOperator` as a static `&str`.
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            UnaryOperator::Neg => "-",
-            UnaryOperator::Not => "!",
-        }
-    }
-}
-
-impl fmt::Display for UnaryOperator {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        self.as_str().fmt(f)
-    }
-}
-
-impl FromStr for UnaryOperator {
-    type Err = Error;
-
-    fn from_str(s: &str) -> Result<Self> {
-        match s {
-            "-" => Ok(UnaryOperator::Neg),
-            "!" => Ok(UnaryOperator::Not),
-            _ => Err(Error::new(format!("invalid unary operator: `{s}`"))),
         }
     }
 }
@@ -202,102 +166,6 @@ fn normalize_rhs(lhs_expr: Expression, operator: BinaryOperator, rhs: BinaryOp) 
     } else {
         // Nothing to normalize.
         BinaryOp::new(lhs_expr, operator, Operation::Binary(rhs))
-    }
-}
-
-/// An operator that can be applied to two expressions.
-#[derive(Debug, PartialEq, Eq, Clone, Copy)]
-pub enum BinaryOperator {
-    /// Equal operator (`==`).
-    Eq,
-    /// Not-equal operator (`!=`).
-    NotEq,
-    /// Less-equal operator (`<=`).
-    LessEq,
-    /// Greater-equal operator (`>=`).
-    GreaterEq,
-    /// Less operator (`<`).
-    Less,
-    /// Greater operator (`>`).
-    Greater,
-    /// Plus operator (`+`).
-    Plus,
-    /// Minus operator (`-`).
-    Minus,
-    /// Multiply operator (`*`).
-    Mul,
-    /// Division operator (`/`).
-    Div,
-    /// Modulo operator (`%`).
-    Mod,
-    /// And operator (`&&`).
-    And,
-    /// Or operator (`||`).
-    Or,
-}
-
-impl BinaryOperator {
-    /// Returns the `BinaryOperator` as a static `&str`.
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            BinaryOperator::Eq => "==",
-            BinaryOperator::NotEq => "!=",
-            BinaryOperator::LessEq => "<=",
-            BinaryOperator::GreaterEq => ">=",
-            BinaryOperator::Less => "<",
-            BinaryOperator::Greater => ">",
-            BinaryOperator::Plus => "+",
-            BinaryOperator::Minus => "-",
-            BinaryOperator::Mul => "*",
-            BinaryOperator::Div => "/",
-            BinaryOperator::Mod => "%",
-            BinaryOperator::And => "&&",
-            BinaryOperator::Or => "||",
-        }
-    }
-
-    // Returns the operator precedence level. Higher numbers mean higher precedence.
-    pub(crate) fn precedence(self) -> u8 {
-        match self {
-            BinaryOperator::Mul | BinaryOperator::Div | BinaryOperator::Mod => 6,
-            BinaryOperator::Plus | BinaryOperator::Minus => 5,
-            BinaryOperator::LessEq
-            | BinaryOperator::GreaterEq
-            | BinaryOperator::Less
-            | BinaryOperator::Greater => 4,
-            BinaryOperator::Eq | BinaryOperator::NotEq => 3,
-            BinaryOperator::And => 2,
-            BinaryOperator::Or => 1,
-        }
-    }
-}
-
-impl fmt::Display for BinaryOperator {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        self.as_str().fmt(f)
-    }
-}
-
-impl FromStr for BinaryOperator {
-    type Err = Error;
-
-    fn from_str(s: &str) -> Result<Self> {
-        match s {
-            "==" => Ok(BinaryOperator::Eq),
-            "!=" => Ok(BinaryOperator::NotEq),
-            "<=" => Ok(BinaryOperator::LessEq),
-            ">=" => Ok(BinaryOperator::GreaterEq),
-            "<" => Ok(BinaryOperator::Less),
-            ">" => Ok(BinaryOperator::Greater),
-            "+" => Ok(BinaryOperator::Plus),
-            "-" => Ok(BinaryOperator::Minus),
-            "*" => Ok(BinaryOperator::Mul),
-            "/" => Ok(BinaryOperator::Div),
-            "%" => Ok(BinaryOperator::Mod),
-            "&&" => Ok(BinaryOperator::And),
-            "||" => Ok(BinaryOperator::Or),
-            _ => Err(Error::new(format!("invalid binary operator: `{s}`"))),
-        }
     }
 }
